changed `undergrad.js` to `year.js`
makes more sense to have all year selections in the same menu & handled by the same file
